+++ examples/middlewares-custom-decorators/decorators/current-user.ts
@@ -1,6 +1,6 @@
-import { createParamDecorator } from "type-graphql";
+import { createParameterDecorator } from "type-graphql";
 import { type Context } from "../context.type";
 
 export function CurrentUser() {
-  return createParamDecorator<Context>(({ context }) => context.currentUser);
+  return createParameterDecorator<Context>(({ context }) => context.currentUser);
 }